Refactor admin views into package
